refactor: organizing folders
