+++ src/domain/LotterySmartContract/entity/Bet.ts
